refactor: extract globals into ivy.transpiler
