+++ ivy/transpiler/transformations/transformers/method_transformer/ivy_method_transformer.py
@@ -28,7 +28,7 @@ from ....utils.ast_utils import (
     MODULE_TO_ALIAS,
     get_function_vars,
 )
-from ... import transformer_globals as glob
+from ivy.transpiler import globals as glob
 
 
 class IvyMethodToFunctionConverter(BaseMethodToFunctionConverter):
